+++ srcs/builtin/builtin_exit.c
@@ -6,11 +6,12 @@
 /*   By: ssoeno <ssoeno@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 14:19:21 by ssoeno            #+#    #+#             */
-/*   Updated: 2024/11/13 21:49:19 by ssoeno           ###   ########.fr       */
+/*   Updated: 2024/11/15 17:24:02 by ssoeno           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
+#include "../../includes/builtin.h"
 
 int	builtin_exit(int argc, char *argv[], t_map *envmap)
 {
